% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueIndex.R
\name{valueIndex}
\alias{valueIndex}
\title{VALUE measure calculation for climate4R grids}
\usage{
valueIndex(grid = NULL, index.code = NULL,
  return.NApercentage = TRUE, parallel = FALSE, max.ncores = 16,
  ncores = NULL)
}
\arguments{
\item{grid}{Grid (also station data) of observations}

\item{index.code}{Characher of the index code to be computed (use VALUE::show.indices).}

\item{return.NApercentage}{Logical to also return or not a grid containing NA percentage information.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A grid of the index or a list containing the grid of the index and the 
grid of NA percenatage
}
\description{
VALUE measure calculation for climate4R grids
}
\examples{
library(transformeR)
y <- EOBS_Iberia_tas
m <- valueIndex(y, "mean")
str(m$Index)
str(m$NApercentage)
}
\author{
M. Iturbide
}
