% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueMeasure.R
\name{valueMeasure}
\alias{valueMeasure}
\title{VALUE measure calculation for climate4R grids}
\usage{
valueMeasure(y, x, measure.code, index.code = NULL,
  return.NApercentage = TRUE, parallel = FALSE, max.ncores = 16,
  ncores = NULL)
}
\arguments{
\item{y}{Grid (also station data) of observations}

\item{x}{Grid (also station data) of the grid that is being validated}

\item{measure.code}{characher of the measure code to be computed (use VALUE::show.measures)}

\item{index.code}{Default is NULL. characher of the index code to be computed (use VALUE::show.indices).}

\item{return.NApercentage}{Logical to also return or not a grid containing NA percentage information.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A grid of the index or a list containing the grid of the index and the 
grid of NA percenatage
}
\description{
VALUE measure calculation for climate4R grids
}
\details{
Some measures are computed directly from the original time series (e.g. temporal correlation),
whereas others are computed upon previouly computed indices (e.g. mean bias). 
Thus, argument \code{index.code} must be provided for the latter case.
}
\author{
M. Iturbide
}
