% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueIndex.R
\name{valueIndex}
\alias{valueIndex}
\title{VALUE measure calculation for climate4R grids}
\usage{
valueIndex(
  grid,
  index.code,
  return.NApercentage = FALSE,
  ...,
  parallel = FALSE,
  max.ncores = 16,
  con = NULL,
  th = NULL,
  which.wetdays = NULL,
  ncores = NULL
)
}
\arguments{
\item{grid}{Grid (also station data) of observations}

\item{index.code}{Character of the index code to be computed (use VALUE::show.indices).}

\item{return.NApercentage}{Logical to also return or not a grid containing NA percentage information.}

\item{...}{Additional arguments pased to FUNCTION.CALL (use VALUE::show.indices to check the available arguments).}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{th}{Numeric value. Threshold used as reference for the condition. Default is NULL. If a threshold value is supplied with no specificaction of the argument \code{th}. Then condition is set to \code{"GE"}.}

\item{which.wetdays}{A string, default to NULL. Infer the measure/index taking into account only the wet days of the temporal serie. 
In this case, set which.wetdays = "Independent".}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}

\item{cond}{Inequality operator to be applied to the given \code{"th"}. Only the days that satisfy the condition will be used for validation the model.
\code{"GT"} = greater than the value of \code{th}, \code{"GE"} = greater or equal,
\code{"LT"} = lower than, \code{"LE"} = lower or equal than.}
}
\value{
A grid of the index or a list containing the grid of the index and the 
grid of NA percenatage
}
\description{
VALUE measure calculation for climate4R grids
}
\examples{
library(transformeR)
require(climate4R.datasets)
data(EOBS_Iberia_tas)
y <- EOBS_Iberia_tas
m <- valueIndex(y, "mean")
str(m$Index)
str(m$NApercentage)
}
\author{
M. Iturbide

M. N. Legasa
}
